test new changees

+++ pconway/core/custom_game.py
@@ -48,6 +48,6 @@ class RandomGame(pconway.core.game.GameOfLife):
             Defaults to 0.
         """
         np.random.seed(seed)
-        matrix = np.zeros((nrow, ncol))
+        matrix = np.zeros((nrow, ncol), dtype=int)
         matrix = pconway.mutation(matrix=matrix, mutation_rate=spawn_rate)
         super().__init__(matrix=matrix, mutation_rate=mutation_rate)
